start implementing reflection and light point.

+++ include/RenderBuffer.h
@@ -1,8 +1,10 @@
 #ifndef __RENDERBUFFER_DECLARATION__
 #define __RENDERBUFFER_DECLARATION__
 
-#include "Vector.h"
 #include <vector>
+#include <SDL2/SDL.h>
+
+#include "Vector.h"
 
 class RenderBuffer
 {
@@ -11,15 +13,15 @@ class RenderBuffer
         RenderBuffer(RenderBuffer const&) = delete;
         ~RenderBuffer() = default;
 
-        auto setColorList(Vector3D<float> col) -> void; 
+        auto setColorList(Vector3D<Uint8> col) -> void; 
         auto setScreenCoordList(Vector2D<float> screen_coord) -> void; 
 
-        auto getColorList() -> std::vector<Vector3D<float>>; 
+        auto getColorList() -> std::vector<Vector3D<Uint8>>; 
         auto getScreenCoordList() -> std::vector<Vector2D<float>>; 
 
     private:
 
-        std::vector<Vector3D<float>> color_list;
+        std::vector<Vector3D<Uint8>> color_list;
         std::vector<Vector2D<float>> screen_coord_list;
 };
