+++ src/RenderBuffer.cpp
@@ -7,7 +7,7 @@ RenderBuffer::RenderBuffer()
     log->info("RenderBuffer created.");
 }
 
-auto RenderBuffer::setColorList(Vector3D<float> col) -> void
+auto RenderBuffer::setColorList(Vector3D<Uint8> col) -> void
 {
     this->color_list.push_back(col);
 }
@@ -17,7 +17,7 @@ auto RenderBuffer::setScreenCoordList(Vector2D<float> screen_coord) -> void
     this->screen_coord_list.push_back(screen_coord);
 }
 
-auto RenderBuffer::getColorList() -> std::vector<Vector3D<float>>
+auto RenderBuffer::getColorList() -> std::vector<Vector3D<Uint8>>
 {
     return this->color_list;
 }
